enquete/requirements.txt
astroid==1.6.1
Django==2.0.2
django-angular==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
inflection==0.3.1
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pylint==1.8.2
pytz==2018.3
six==1.11.0
wrapt==1.10.11
